backend/requirements.txt
flask==2.0.1
face-recognition==1.3.0
numpy==1.21.2
opencv-python==4.5.3.56
pillow==8.3.1
python-dotenv==0.19.0
mysql-connector-python==8.0.27
Werkzeug==2.0.1
pytest==7.4.3
hypothesis==6.92.1
qrcode==7.4.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
